OW-576,OW-614,implemented 614, passed testson Java for new models

+++ rscripts/JavaLpSolve.R
@@ -18,7 +18,7 @@ CallLpSolve <- function(lpObj_vec,lpCon_mat,lpDir_vec,lpRhs_vec,
   lpModel <- LpSolve$makeLp(0L, varNum)
 
 #  if(!missing(verbose)){
-  lpModel$setVerbose(as.integer(4));
+  lpModel$setVerbose(as.integer(3));
 #  }
 
   # set objective
